swaggerの設定を別ファイルに移動
