+++ lib/src/index.ts
@@ -2,7 +2,6 @@ import 'reflect-metadata';
 
 import serverlessExpress from '@codegenie/serverless-express';
 import express from 'express';
-import swaggerJSDoc from 'swagger-jsdoc';
 import swaggerUi from 'swagger-ui-express';
 import { container } from 'tsyringe';
 
@@ -36,6 +35,7 @@ import type { IRaceRepository } from './repository/interface/IRaceRepository';
 import { GoogleCalendarService } from './service/implement/googleCalendarService';
 import type { ICalendarService } from './service/interface/ICalendarService';
 import { MockGoogleCalendarService } from './service/mock/mockGoogleCalendarService';
+import swaggerSpec from './swagger/swaggerConfig';
 import { JraPlaceDataUseCase } from './usecase/implement/jraPlaceDataUseCase';
 import { JraRaceCalendarUseCase } from './usecase/implement/jraRaceCalendarUseCase';
 import { JraRaceDataUseCase } from './usecase/implement/jraRaceDataUseCase';
@@ -284,25 +284,9 @@ app.use(express.json());
 app.use('/api/races/nar', narRaceController.router);
 app.use('/api/races/jra', jraRaceController.router);
 
-const swaggerDefinition = {
-    basePath: '/',
-    openapi: '3.0.0',
-    info: {
-        version: '1.0.0',
-        title: 'My API',
-        description: 'Loading and retrieving  data from firebase firestore',
-    },
-    servers: [{ url: 'http://localhost:3000/' }],
-};
-
-const swaggerOptions = {
-    swaggerDefinition,
-    apis: ['./**/*.ts'],
-};
-const swaggerDocs = swaggerJSDoc(swaggerOptions);
-
+// Swaggerの設定
 app.use('/api-docs', swaggerUi.serve);
-app.get('/api-docs', swaggerUi.setup(swaggerDocs, { explorer: true }));
+app.get('/api-docs', swaggerUi.setup(swaggerSpec, { explorer: true }));
 
 // Lambda用のハンドラーをエクスポート
 export const handler = serverlessExpress({ app });
